+++ tw-entrypoints/src/main/java/com/transferwise/common/entrypoints/tableaccessstatistics/TasQueryParsingListener.java
@@ -1,0 +1,10 @@
+package com.transferwise.common.entrypoints.tableaccessstatistics;
+
+import java.time.Duration;
+
+public interface TasQueryParsingListener {
+
+  void parsingDone(String sql, ParsedQuery parsedQuery, Duration timeTaken);
+
+  void parsingFailed(String sql, Duration timeTaken, Throwable t);
+}